Add support for configuration files
The config.js file will not be commited to the public repo, and
will be used to store secrets like sendgrid API keys.

Currently this will copy config.default.js to config.js if no
config file exists, but another option might be to load both
config.default.js AND default.js, and combine them. This would
prevent maintenance if new options are added and prevent errors
if options don't exist, but would make it difficult to notice if
a config file mistakenly doesn't exist.